// src/app/models/product-get-list-request-payload.ts
export class ProductGetListRequestPayload {
    StoreId: number;
    AppId: number;
    PageSize: number;
    PageNumber: number;
    IsClub: number;
    CategoryId: string;
    SessionId: string;
    UserId: number;
    ListId: string;
    SizeId: string;
    TypeId: string;
    VarietalId: string;
    CountryId: string;
    RegionId: string;
    IsFavorite: number;
    IsFeatureProduct: number;
    IsStaffPicks: number;
   PriceRangeId: string;
    KeyWord: string;
    DeviceId: string;
    DeviceType: string;
  Review: number;
    SortBy: number;
}